Move histopyramid files to gpgpu module
