+++ lerna.json
@@ -11,7 +11,6 @@
       "ignore": [
         "test",
         "*.md",
-        "@luma.gl/gpgpu",
         "@luma.gl/imageprocessing",
         "@luma.gl/io",
         "@luma.gl/scripts"
