+++ modules/gpgpu/README.md
@@ -1,0 +1,3 @@
+# @luma.gl/gpgpu
+
+This module is a collection of several utility functions that are accelerated by the GPU.
