+++ modules/gpgpu/src/histopyramid/histopyramid.md
