+++ modules/gpgpu/src/index.js
@@ -1,3 +1,9 @@
 // GPGPU utilities for luma.gl
 
 export {diffImagePixels} from './diff-images';
+
+export {
+  buildHistopyramidBaseLevel as _buildHistopyramidBaseLevel,
+  getHistoPyramid as _getHistoPyramid,
+  histoPyramidGenerateIndices as _histoPyramidGenerateIndices
+} from './histopyramid/histopyramid';
